Fixed int/float conversion error.
Former-commit-id: c5efc3ecf594b702ecb7110e46aa0d4c9920d703
Former-commit-id: e68d718ef8d3e87339fc0782f572737abc9a2b53
Former-commit-id: 39d2094e414d3ef205d9d83522fb293e56aae977
Former-commit-id: d52bf1d6974afaed810d22a888c78d4364cee933
Former-commit-id: d1d054fc7e7c706eda9b7881feb8b0c955d68b44
Former-commit-id: eea263945101e0783c587d7e099a9c206d7c20eb

+++ dbscripts/migrate_bookratingcsv_to_title.py
@@ -37,7 +37,7 @@ def importRatings():
         
         if (foundTitle != None):
             # Update the title object in the database based on the current rating row
-            if int(row['Overall']) >= 3:
+            if float(row['Overall']) >= 3.0:
                 foundTitle.promoter_count = foundTitle.promoter_count + 1
             else:
                 foundTitle.detractor_count = foundTitle.detractor_count + 1
